BE refactor model APIs, FE bug fixes
