+++ crypticsage/src/api/adminController.js
@@ -178,21 +178,13 @@ export const executeTalibFunction = async ({ token, payload }) => {
     return response;
 }
 
-export const getCorelationMatrix = async ({ token, payload }) => {
-    const config = {
-        headers: { Authorization: `Bearer ${token}` }
-    };
-    const response = await axios.post(`${baseUrl}/indicators/get_corelation_matrix`, { payload }, config, {
-        withCredentials: true
-    })
-    return response;
-}
+// <---------------------- Model related APIs ---------------------->
 
 export const startModelTraining = async ({ token, payload }) => {
     const config = {
         headers: { Authorization: `Bearer ${token}` }
     };
-    const response = await axios.post(`${baseUrl}/indicators/start_model_training`, { payload }, config, {
+    const response = await axios.post(`${baseUrl}/model/start_model_training`, { payload }, config, {
         withCredentials: true
     })
     return response;
@@ -202,7 +194,17 @@ export const retrain_wgan_Model = async ({ token, payload }) => {
     const config = {
         headers: { Authorization: `Bearer ${token}` }
     };
-    const response = await axios.post(`${baseUrl}/indicators/retrain_model`, { payload }, config, {
+    const response = await axios.post(`${baseUrl}/model/retrain_model`, { payload }, config, {
+        withCredentials: true
+    })
+    return response;
+}
+
+export const getCorelationMatrix = async ({ token, payload }) => {
+    const config = {
+        headers: { Authorization: `Bearer ${token}` }
+    };
+    const response = await axios.post(`${baseUrl}/model/get_corelation_matrix`, { payload }, config, {
         withCredentials: true
     })
     return response;
@@ -212,7 +214,7 @@ export const getUserModels = async ({ token }) => {
     const config = {
         headers: { Authorization: `Bearer ${token}` }
     };
-    const response = await axios.post(`${baseUrl}/indicators/get_model`, { withCredentials: true }, config)
+    const response = await axios.post(`${baseUrl}/model/get_model`, { withCredentials: true }, config)
     return response;
 }
 
@@ -220,17 +222,7 @@ export const saveModel = async ({ token, payload }) => {
     const config = {
         headers: { Authorization: `Bearer ${token}` }
     };
-    const response = await axios.post(`${baseUrl}/indicators/save_model`, { payload }, config, {
-        withCredentials: true
-    })
-    return response;
-}
-
-export const deleteModelForUser = async ({ token, model_id, model_type }) => {
-    const config = {
-        headers: { Authorization: `Bearer ${token}` }
-    };
-    const response = await axios.post(`${baseUrl}/indicators/delete_user_model`, { model_id, model_type }, config, {
+    const response = await axios.post(`${baseUrl}/model/save_model`, { payload }, config, {
         withCredentials: true
     })
     return response;
@@ -241,55 +233,28 @@ export const deleteModel = async ({ token, payload }) => {
         headers: { Authorization: `Bearer ${token}` }
     };
     const { model_id, model_type, asset_type, ticker_name, period } = payload;
-    const response = await axios.post(`${baseUrl}/indicators/delete_model`, { model_id, model_type, asset_type, ticker_name, period }, config, {
+    const response = await axios.post(`${baseUrl}/model/delete_model`, { model_id, model_type, asset_type, ticker_name, period }, config, {
         withCredentials: true
     })
     return response;
 }
 
-// Not being used anymore
-export const checkIfModelDataExists = async ({ token, payload }) => {
+export const deleteModelForUser = async ({ token, model_id, model_type }) => {
     const config = {
         headers: { Authorization: `Bearer ${token}` }
     };
-    const { model_id, modelType } = payload;
-    const response = await axios.post(`${baseUrl}/indicators/check_for_model`, { model_id, modelType }, config, {
+    const response = await axios.post(`${baseUrl}/model/delete_user_model`, { model_id, model_type }, config, {
         withCredentials: true
     })
     return response;
 }
 
-// Not being used anymore
-export const getModelResult = async ({ token, payload }) => {
+export const addNewTrainingResults = async ({ token, payload }) => {
     const config = {
-        headers: { Authorization: `Bearer ${token}` }
+        headers: { "Authorization": `Bearer ${token}` },
     };
-    const { model_id } = payload;
-    const response = await axios.post(`${baseUrl}/indicators/get_model_result`, { model_id }, config, {
-        withCredentials: true
-    })
+    const response = await axios.post(`${baseUrl}/model/update_new_predictions`, { payload }, config)
     return response;
-}
-
-export const makeNewPrediction = async ({ token, payload }) => {
-    const config = {
-        headers: { Authorization: `Bearer ${token}` }
-    };
-    const response = await axios.post(`${baseUrl}/indicators/make_new_prediction`, { payload }, config, {
-        withCredentials: true
-    })
-    return response;
-}
-
-export const makeWganPrediction = async ({ token, payload }) => {
-    const config = {
-        headers: { Authorization: `Bearer ${token}` }
-    };
-    const response = await axios.post(`${baseUrl}/indicators/make_wgan_prediction`, { payload }, config, {
-        withCredentials: true
-    })
-    return response;
-
 }
 
 export const renameModel = async ({ token, payload }) => {
@@ -297,7 +262,7 @@ export const renameModel = async ({ token, payload }) => {
         headers: { Authorization: `Bearer ${token}` }
     };
     const { model_id, model_name } = payload;
-    const response = await axios.post(`${baseUrl}/indicators/rename_model`, { model_id, model_name }, config, {
+    const response = await axios.post(`${baseUrl}/model/rename_model`, { model_id, model_name }, config, {
         withCredentials: true
     })
     return response;
@@ -308,14 +273,27 @@ export const getModelCheckPoints = async ({ token, model_id }) => {
         headers: { "Authorization": `Bearer ${token}` },
         withCredentials: true
     };
-    const response = await axios.get(`${baseUrl}/indicators/get_model_checkpoints?model_id=${model_id}`, config)
+    const response = await axios.get(`${baseUrl}/model/get_model_checkpoints?model_id=${model_id}`, config)
     return response;
 }
 
-export const addNewTrainingResults = async ({ token, payload }) => {
+export const makeNewPrediction = async ({ token, payload }) => {
     const config = {
-        headers: { "Authorization": `Bearer ${token}` },
+        headers: { Authorization: `Bearer ${token}` }
     };
-    const response = await axios.post(`${baseUrl}/indicators/update_new_predictions`, { payload }, config)
+    const response = await axios.post(`${baseUrl}/model/make_new_prediction`, { payload }, config, {
+        withCredentials: true
+    })
     return response;
+}
+
+export const makeWganPrediction = async ({ token, payload }) => {
+    const config = {
+        headers: { Authorization: `Bearer ${token}` }
+    };
+    const response = await axios.post(`${baseUrl}/model/make_wgan_prediction`, { payload }, config, {
+        withCredentials: true
+    })
+    return response;
+
 }
